requirements.txt
cassandra-driver==3.26.0
click==8.1.3
geomet==0.2.1.post1
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
